Fix 2015 tests

+++ 2015/src/test/java/Day1Test.java
@@ -10,11 +10,11 @@ public class Day1Test {
 
     @Test
     public void testPart1Output() {
-        assertEquals(138, day.part1());
+        assertEquals(138, (int) day.part1());
     }
 
     @Test
     public void testPart2Output() {
-        assertEquals(1771, day.part2());
+        assertEquals(1771, (int) day.part2());
     }
 }
